Create FINAL CODE - Trump Eats Money and Destroys the Environment (the Withdrawal from the Paris Agreement)
I built on the original code that we used by changing the shapes of the goals and the player, adding a backstory, created a scoring system(three scores each having different meanings and also gave them colour), and made a decrease in speed. I also removed the sounds because they made the game lag.